backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-api-python-client==2.97.0
pydantic==2.3.0
python-dotenv==1.0.0
